# Optimize `r_lgl_sum()` and `r_lgl_which()`