@DavisVaughan 2023-02-24
I checked coverage of that file and all lines of `r_lgl_sum()` and `r_lgl_which()` were already being hit by tests, so that was good. But I've added some more edge case ones anyways!